django/requirements.txt
django-rest-framework
djangorestframework-simplejwt
awscli
boto3
psycopg2